\name{esercizio.1.4.10}
\alias{esercizio.1.4.10}
\docType{data}
\title{Data frame Esercizio 1.4.10}
\description{
  Studio sulla produttività di un ortaggio coltivato in una serra.
}
\usage{data(esercizio.1.4.10)}
\format{
  Data frame composto di 16 osservazioni nelle seguenti 4 variabili:
  \describe{
    \item{\code{Ustatistica} }{unità statistica}
    \item{\code{Varieta} }{tipo di varietà}
    \item{\code{Raccolto} }{raccolto}
    \item{\code{Commestibile} }{parte commestibile}
  }
}
\references{
 Stefanini, F. M.
 \emph{INTRODUZIONE ALLA STATISTICA APPLICATA con esempi in R.}
 PEARSON Education Milano, 2007, \url{http://hpe.pearsoned.it/stefanini}.
}
\examples{
data(esercizio.1.4.10)
str(esercizio.1.4.10)
}
\keyword{datasets}
