\name{esercizio.1.4.14}
\alias{esercizio.1.4.14}
\docType{data}
\title{Data frame Esercizio 1.4.14}
\description{
  L'azienda Pinewood produce manici di badile in legno. Le variabili diametro, lunghezza e peso
  sono state misurate in un campione di manici prodotto con legno differente.
}
\usage{data(esercizio.1.4.14)}
\format{
  Data frame composto di 15 osservazioni nelle seguenti 3 variabili:
  \describe{
    \item{\code{Diametro}}{diametro}
    \item{\code{Lunghezza}}{lunghezza}
    \item{\code{Peso}}{peso}
  }
}
\references{
 Stefanini, F. M.
 \emph{INTRODUZIONE ALLA STATISTICA APPLICATA con esempi in R.}
 PEARSON Education Milano, 2007, \url{http://hpe.pearsoned.it/stefanini}.
}
\examples{
data(esercizio.1.4.14)
str(esercizio.1.4.14)
}
\keyword{datasets}
