\name{esercizio.4.4.1}
\alias{esercizio.4.4.1}
\docType{data}
\title{Data frame Esercizio 4.4.1}
\description{
  In uno studio inerente malattie bovine è stato studiato un campione
  casuale di 301 animali. La distribuzione congiunta delle frequenze assolute
  per le variabili malattia (Y = 1 se presente, Y = 0 se assente) e età (X = 0 per
  12 - 18 mesi, X = 1 per 19 - 24 mesi, X = 2 per oltre 24 mesi) è rappresentata
  nel presente data frame.
}
\usage{data(esercizio.4.4.1)}
\format{
  Data frame composto di 6 osservazioni nelle seguenti 3 variabili:
  \describe{
    \item{\code{Y}}{malattia}
    \item{\code{X}}{età}
    \item{\code{frequenze}}{frequenze assolute}
  }
}
\references{
 Stefanini, F. M.
 \emph{INTRODUZIONE ALLA STATISTICA APPLICATA con esempi in R.}
 PEARSON Education Milano, 2007, \url{http://hpe.pearsoned.it/stefanini}.
}
\examples{
data(esercizio.4.4.1)
str(esercizio.4.4.1)
}
\keyword{datasets}

