\name{esercizio.5.3.10}
\alias{esercizio.5.3.10}
\docType{data}
\title{Data frame Esercizio 5.3.10}
\description{
  In uno studio sulla regolazione dei dispensatori di bevande calde viene esaminato
  un campione casuale di 16 bicchierini ottenuti da 4 dispensatori. Per ogni erogazione
  si misura il tempo impiegato per riempire il bicchiere e la temperatura al termine
  dell'erogazione.
}
\usage{data(esercizio.5.3.10)}
\format{
  Data frame composto di 16 osservazioni nelle seguenti 3 variabili:
  \describe{
    \item{\code{secondi}}{tempo}
    \item{\code{tempera}}{temperatura}
    \item{\code{dispensatore}}{dispensatore}
  }
}
\references{
 Stefanini, F. M.
 \emph{INTRODUZIONE ALLA STATISTICA APPLICATA con esempi in R.}
 PEARSON Education Milano, 2007, \url{http://hpe.pearsoned.it/stefanini}.
}
\examples{
data(esercizio.5.3.10)
str(esercizio.5.3.10)
}
\keyword{datasets}
