\name{esercizio.5.3.11}
\alias{esercizio.5.3.11}
\docType{data}
\title{Data frame Esercizio 5.3.11}
\description{
  In uno studio sulla concentrazione Y di polifenoli in bottiglie di vino è stata misurata la concentrazione
  di flavonosimili X prima della fermentazione (file \dQuote{eseC5.011.R}).
}
\usage{data(esercizio.5.3.11)}
\format{
  Data frame composto di 24 osservazioni nelle seguenti 2 variabili:
  \describe{
    \item{\code{flavo}}{concentrazione di flavonosimili}
    \item{\code{polifenoli}}{concentrazione di polifenoli}
  }
}
\references{
 Stefanini, F. M.
 \emph{INTRODUZIONE ALLA STATISTICA APPLICATA con esempi in R.}
 PEARSON Education Milano, 2007, \url{http://hpe.pearsoned.it/stefanini}.
}
\examples{
data(esercizio.5.3.11)
str(esercizio.5.3.11)
}
\keyword{datasets}
