\name{esercizio.5.3.14}
\alias{esercizio.5.3.14}
\docType{data}
\title{Data frame Esercizio 5.3.14}
\description{
  Nel file \dQuote{eseC5.014.R} è contenuto il dataframe minestrone che contiene le variabili fibra,
  grass, carboidrat, caloria per indicare rispettivamente le quantità di fibra, grassi, di carboidrati
  e le calorie misurate su 227 unità statistiche costituite da porzioni standard di minestrone congelato
  proveniente da buste casualmente estratte dalla produzione di gennaio 2006.
}
\usage{data(esercizio.5.3.14)}
\format{
  Data frame composto di 227 osservazioni nelle seguenti 4 variabili:
  \describe{
    \item{\code{fibra}}{quantità di fibra}
    \item{\code{grass}}{quantità di grassi}
    \item{\code{carboidrat}}{quantità di carboidrati}
    \item{\code{caloria}}{quantità di calorie}
  }
}
\references{
 Stefanini, F. M.
 \emph{INTRODUZIONE ALLA STATISTICA APPLICATA con esempi in R.}
 PEARSON Education Milano, 2007, \url{http://hpe.pearsoned.it/stefanini}.
}
\examples{
data(esercizio.5.3.14)
str(esercizio.5.3.14)
}
\keyword{datasets}
