\name{esercizio.5.3.15}
\alias{esercizio.5.3.15}
\docType{data}
\title{Data frame Esercizio 5.3.15}
\description{
  In uno studio inerente un processo fermentativo si considera la concentrazione
  finale di polifenoli Y in relazione alla concentrazione iniziale di zuccheri X
  e alla tipologia di ceppo di lievito Z. I dati sono contenuti nel file \dQuote{esec5.015.R}.
}
\usage{data(esercizio.5.3.15)}
\format{
  Data frame composto di 200 osservazioni nelle seguenti 3 variabili:
  \describe{
    \item{\code{Y}}{concentrazione finale di polifenoli}
    \item{\code{X}}{concentrazione iniziale di zuccheri}
    \item{\code{Z}}{tipologia di ceppo di lievito}
  }
}
\references{
 Stefanini, F. M.
 \emph{INTRODUZIONE ALLA STATISTICA APPLICATA con esempi in R.}
 PEARSON Education Milano, 2007, \url{http://hpe.pearsoned.it/stefanini}.
}
\examples{
data(esercizio.5.3.15)
str(esercizio.5.3.15)
}
\keyword{datasets}
