\name{esercizio.5.3.5}
\alias{esercizio.5.3.5}
\docType{data}
\title{Data frame Esercizio 5.3.5}
\description{
  In un esperimento di confronto è stato misurato il peso dell'apparato radicale in un campione di 5 piante, prima
  con un metodo meccanico (variabile X), poi con un metodo elettronico (variabile Y). 
}
\usage{data(esercizio.5.3.5)}
\format{
  Data frame composto di 5 osservazioni nelle seguenti 2 variabili:
  \describe{
    \item{\code{y}}{metodo elettrico}
    \item{\code{x}}{metodo meccanico}
  }
}
\references{
 Stefanini, F. M.
 \emph{INTRODUZIONE ALLA STATISTICA APPLICATA con esempi in R.}
 PEARSON Education Milano, 2007, \url{http://hpe.pearsoned.it/stefanini}.
}
\examples{
data(esercizio.5.3.5)
str(esercizio.5.3.5)
}
\keyword{datasets}
