\name{esercizio.5.3.6}
\alias{esercizio.5.3.6}
\docType{data}
\title{Data frame Esercizio 5.3.6}
\description{
  In un esperimento sulla qualità del latte è stato contato il numero di batteri (scala logaritmica)
  presenti in un campione di 7 provette che inizialmente hanno lo stesso titolo batterico e che sono
  analizzate dopo trattamento termico a temperature differenti.
}
\usage{data(esercizio.5.3.6)}
\format{
  Data frame composto di 7 osservazioni nelle seguenti 2 variabili:
  \describe{
    \item{\code{tempera}}{temperatura}
    \item{\code{titolo}}{titolo batterico}
  }
}
\references{
 Stefanini, F. M.
 \emph{INTRODUZIONE ALLA STATISTICA APPLICATA con esempi in R.}
 PEARSON Education Milano, 2007, \url{http://hpe.pearsoned.it/stefanini}.
}
\examples{
data(esercizio.5.3.6)
str(esercizio.5.3.6)
}
\keyword{datasets}
