\name{esercizio.5.3.9}
\alias{esercizio.5.3.9}
\docType{data}
\title{Data frame Esercizio 5.3.9}
\description{
  In uno studio inerente la concentrazione di ferro X nel sangue di individui della
  Cinta Senese, sono stati raccolti 3 campioni casuali in 3 tempi successivi (variabile tempo).
}
\usage{data(esercizio.5.3.9)}
\format{
  Data frame composto di 300 osservazioni nelle seguenti 2 variabili:
  \describe{
    \item{\code{tempo}}{tempo}
    \item{\code{X}}{concentrazione di ferro}
  }
}
\references{
 Stefanini, F. M.
 \emph{INTRODUZIONE ALLA STATISTICA APPLICATA con esempi in R.}
 PEARSON Education Milano, 2007, \url{http://hpe.pearsoned.it/stefanini}.
}
\examples{
data(esercizio.5.3.9)
str(esercizio.5.3.9)
}
\keyword{datasets}
