\name{ISA-package}
\alias{ISA-package}
\docType{package}
\title{
INTRODUZIONE ALLA STATISTICA APPLICATA con esempi in R
}
\description{
Insieme di funzioni di supporto al volume "INTRODUZIONE
ALLA STATISTICA APPLICATA con esempi in R", Federico M.
Stefanini, PEARSON Education Milano, 2007.
}
\details{
\tabular{ll}{
Package: \tab ISA\cr
Type: \tab Package\cr
LazyData: \tab no\cr
Encoding: \tab latin1\cr
Depends: \tab R (>= 2.4.1), stats, utils, tools\cr
Version: \tab 1.0-32\cr
Date: \tab 2007-09-14\cr
License: \tab GPL-2\cr
URL: \tab http://hpe.pearsoned.it/stefanini\cr
}

Index:
\preformatted{
Funzioni:
=========

boxPstd                 Boxplot standard
coefVariaz              Coefficiente di variazione
difSempliceMed          Differenza semplice media
fiveNumbersSummary      Cinque numeri di Tukey
FLinearMod              Minimi quadrati ordinari
media                   Media aritmetica
moda                    Moda
quantiB                 Quantili
varpopol                Varianza nella popolazione


Data frames:
============

esercizio.1.4.10        Data frame Esercizio 1.4.10
esercizio.1.4.14        Data frame Esercizio 1.4.14
esercizio.4.4.1         Data frame Esercizio 4.4.1
esercizio.4.4.13        Data frame Esercizio 4.4.13
esercizio.4.4.14        Data frame Esercizio 4.4.14
esercizio.5.3.1         Data frame Esercizio 5.3.1
esercizio.5.3.3         Data frame Esercizio 5.3.3
esercizio.5.3.5         Data frame Esercizio 5.3.5
esercizio.5.3.6         Data frame Esercizio 5.3.6
esercizio.5.3.7         Data frame Esercizio 5.3.7
esercizio.5.3.8         Data frame Esercizio 5.3.8
esercizio.5.3.9         Data frame Esercizio 5.3.9
esercizio.5.3.10        Data frame Esercizio 5.3.10
esercizio.5.3.11        Data frame Esercizio 5.3.11
esercizio.5.3.12        Data frame Esercizio 5.3.12
esercizio.5.3.14        Data frame Esercizio 5.3.14
esercizio.5.3.15        Data frame Esercizio 5.3.15
esercizio.5.3.16        Data frame Esercizio 5.3.16
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{SoluzioniCapitolo1} \tab Soluzioni degli esercizi di ricapitolazione Capitolo 1 (source, pdf)\cr
\code{SoluzioniCapitolo2} \tab Soluzioni degli esercizi di ricapitolazione Capitolo 2 (source, pdf)\cr
\code{SoluzioniCapitolo3} \tab Soluzioni degli esercizi di ricapitolazione Capitolo 3 (source, pdf)\cr
\code{SoluzioniCapitolo4} \tab Soluzioni degli esercizi di ricapitolazione Capitolo 4 (source, pdf)\cr
\code{SoluzioniCapitolo5} \tab Soluzioni degli esercizi di ricapitolazione Capitolo 5 (source, pdf)\cr
}
}
\author{
Fabio Frascati <fabiofrascati@yahoo.it> and Federico M.
Stefanini <stefanini@ds.unifi.it>

Maintainer: Fabio Frascati <fabiofrascati@yahoo.it>
}
\keyword{ package }
